update linux version

+++ trinity/src/trinity_widget_surface.cpp
@@ -7,7 +7,7 @@ namespace NextAI
 
 WidgetSurface::WidgetSurface(ObjectId id): WidgetObject(id)
 {
-	Rect full = { 0, 0, APP_SERVICE()->getWindowsWidth(), APP_SERVICE()->getWindowsHeight() };
+	Rect full = { 0, 0, (int32)APP_SERVICE()->getWindowsWidth(), (int32)APP_SERVICE()->getWindowsHeight() };
 	setArea(full);
 }
